More fixes for PowerPC GCC (the konamiic ones are actual errors!)

+++ src/mame/machine/atari.c
@@ -324,7 +324,7 @@ static void console_write(const address_space *space, UINT8 data)
 {
 	const device_config *dac = devtag_get_device(space->machine, "dac");
 	if (data & 0x08)
-		dac_data_w(dac, -120);
+		dac_data_w(dac, (UINT8)-120);
 	else
 		dac_data_w(dac, +120);
 }
