+++ src/mame/video/fastlane.c
@@ -174,7 +174,7 @@ VIDEO_UPDATE( fastlane )
 	tilemap_set_scrolly(state->layer0, 0, k007121_ctrlram_r(state->k007121, 2));
 
 	tilemap_draw(bitmap, &finalclip0, state->layer0, 0, 0);
-	k007121_sprites_draw(state->k007121, bitmap, cliprect, screen->machine->gfx[0], screen->machine->colortable, state->spriteram, 0, 40, 0, -1);
+	k007121_sprites_draw(state->k007121, bitmap, cliprect, screen->machine->gfx[0], screen->machine->colortable, state->spriteram, 0, 40, 0, (UINT32)-1);
 	tilemap_draw(bitmap, &finalclip1, state->layer1, 0, 0);
 	return 0;
 }
