+++ src/mame/video/flkatck.c
@@ -170,7 +170,7 @@ VIDEO_UPDATE( flkatck )
 
 	/* draw the graphics */
 	tilemap_draw(bitmap, &clip[0], state->k007121_tilemap[0], 0, 0);
-	k007121_sprites_draw(state->k007121, bitmap, cliprect, screen->machine->gfx[0], NULL, &state->k007121_ram[0x1000], 0, 40, 0, -1);
+	k007121_sprites_draw(state->k007121, bitmap, cliprect, screen->machine->gfx[0], NULL, &state->k007121_ram[0x1000], 0, 40, 0, (UINT32)-1);
 	tilemap_draw(bitmap, &clip[1], state->k007121_tilemap[1], 0, 0);
 	return 0;
 }
